feat: release version 1.0.0 with major updates and refactoring
- Updated version number in composer.json to 1.0.0.
- Refactored kui-accordion block to use a single default template, removing old variants.
- Added new kui-accordion_default.twig for default accordion behavior.
- Removed unused kui-accordion_simple.twig and kui-accordion_rich.twig files.
- Updated various block templates (carousel, dialog, drawer, select, tabs) to support project-level overrides.
- Refactored SCSS files to use CSS custom properties for spacing, borders, and transitions.
- Deleted _tokens.scss as design tokens are now defined in _root.scss.
- Improved documentation in block templates for clarity on usage and customization.

+++ snippets/ui-library/assets.php
@@ -19,7 +19,7 @@
 //   {{ snippet('kui/assets', ['js' => false]) | raw }}
 $js = $js ?? true;
 
-$port      = option('kui.components.vite.port', 5174);
+$port      = (int) option('kui.components.vite.port', 5174);
 $devOrigin = 'http://localhost:' . $port;
 
 if (option('debug')) {
